Rename ambiguous header
