+++ winml/lib/Common/inc/cppwinrt_onnx.h
@@ -3,5 +3,5 @@
 
 #pragma once
 
-#include "cppwinrt.h"
+#include "winrt_headers.h"
 #include "onnx.h"
